+++ Sources/StreamFeeds/Models/BookmarkData.swift
@@ -6,7 +6,7 @@ import Foundation
 import StreamCore
 
 public struct BookmarkData: Equatable, Sendable {
-    public let activity: ActivityData
+    public private(set) var activity: ActivityData
     public let createdAt: Date
     public let custom: [String: RawJSON]?
     public internal(set) var folder: BookmarkFolderData?
@@ -17,6 +17,26 @@ public struct BookmarkData: Equatable, Sendable {
 extension BookmarkData: Identifiable {
     public var id: String {
         "\(user.id)-\(activity.id)"
+    }
+}
+
+// MARK: - Mutating the Data
+
+extension BookmarkData {
+    mutating func merge(with incomingData: ActivityData) {
+        activity.merge(with: incomingData)
+    }
+    
+    // MARK: - Current Feed Capabilities
+    
+    mutating func mergeFeedOwnCapabilities(from capabilitiesMap: [FeedId: Set<FeedOwnCapability>]) {
+        activity.mergeFeedOwnCapabilities(from: capabilitiesMap)
+    }
+    
+    func withFeedOwnCapabilities(from capabilitiesMap: [FeedId: Set<FeedOwnCapability>]) -> BookmarkData {
+        var updated = self
+        updated.mergeFeedOwnCapabilities(from: capabilitiesMap)
+        return updated
     }
 }